[refactor♻️] api 엔드 포인트 통일 및 User DTO 디렉토리 구조 변경
